Breaking out files for csharp
